requirements.txt
# Core dependencies
click>=8.0.0
pyyaml>=6.0
python-dotenv>=0.19.0

# Database drivers
psycopg2-binary>=2.9.0
pymongo>=4.0.0

# Storage
boto3>=1.20.0

# Utilities
rich>=12.0.0

# Development dependencies (add with pip install -r requirements-dev.txt)
# pytest>=7.0.0
# black>=22.0.0
# flake8>=4.0.0
# mypy>=0.950
